Refactor module:git with jsdoc and tests
These changes use a test-driven and documentation-driven approach
to refactoring the git module. The focus was on upholding most of the
existing API while iterating on the module's internals. However, some
changes to the API were necessary. These changes have been
highlighted below.

The following changes occurred to the module's structure:

- The `git.is` method has been renamed to `git.isAvailable`. Its logic
was also rewritten for simplicity, clarity, and testability.

- The `git.isPackage` method has been added. The method uses
`git rev-parse` with the `-C` flag to confirm whether the given path
points to a Git repository.

- While the `git` object is exported by default, these changes also
export the constant, allowing devs to import the module as a named
export, e.g. `import { git } from './git.js'`.